+++ packages/isaacscript-common/src/features/characterStats.ts
@@ -10,8 +10,8 @@ import {
   PlayerType,
 } from "isaac-typescript-definitions";
 import { errorIfFeaturesNotInitialized } from "../featuresInitialized";
-import { getDefaultPlayerStat } from "../functions/cacheFlag";
 import { addStat } from "../functions/players";
+import { getDefaultPlayerStat } from "../functions/stats";
 
 const FEATURE_NAME = "characterStats";